* 增强hash能力

+++ langx-java/src/main/java/com/jn/langx/util/hash/Crc32cHasher.java
@@ -10,7 +10,7 @@ public class Crc32cHasher extends Hasher {
     }
 
     @Override
-    public int get() {
+    public long get() {
         reset();
         return crc;
     }
